yt/requirements.txt
crewai
crewai-tools
load_dotenv
langchain-groq
pydantic==2.10.0